Neaten and refactor. Start of dealing with contentid issue, but will require a schema update too
